fedocallib: RRULE INTERVAL must be an integer
Per RFC5545:
> The INTERVAL rule part contains a positive integer representing at
> which intervals the recurrence rule repeats.

Having INTERVAL be a decimal breaks import into some calendars,
like Google's.

Fixes: https://pagure.io/fedocal/issue/203

+++ fedocal/fedocallib/__init__.py
@@ -705,7 +705,7 @@ def add_meeting_to_vcal(ical, meeting, reminder=None):
 
     if meeting.recursion_frequency and meeting.recursion_ends:
         newrule = rrule.rruleset()
-        freq = meeting.recursion_frequency / 7
+        freq = int(meeting.recursion_frequency / 7)
         recursion_ends = datetime.combine(
             meeting.recursion_ends, time(0)).replace(tzinfo=pytz.utc)
         newrule.rrule(
